feat?(api): add get data buffer utils
