+++ api/src/prescription/learner-management/domain/usecases/import-from-feature/send-organization-learners-file.js
@@ -1,6 +1,7 @@
 import { createReadStream } from 'fs';
 
 import { CommonCsvLearnerParser } from '../../../infrastructure/serializers/csv/common-csv-learner-parser.js';
+import { getDataBuffer } from '../../../infrastructure/utils/bufferize/get-data-buffer.js';
 import { AggregateImportError, OrganizationLearnerImportFormatNotFoundError } from '../../errors.js';
 import { OrganizationImport } from '../../models/OrganizationImport.js';
 
@@ -52,18 +53,5 @@ const sendOrganizationLearnersFile = async function ({
     }
   }
 };
-
-function getDataBuffer(readableStream) {
-  return new Promise((resolve, reject) => {
-    const chunks = [];
-    readableStream.on('data', (data) => {
-      chunks.push(data);
-    });
-    readableStream.on('error', (err) => reject(err));
-    readableStream.once('end', () => {
-      resolve(Buffer.concat(chunks));
-    });
-  });
-}
 
 export { sendOrganizationLearnersFile };
