+++ api/src/prescription/learner-management/domain/usecases/replace-sup-organization-learner.js
@@ -1,6 +1,7 @@
 import { createReadStream } from 'fs';
 
 import { SupOrganizationLearnerParser } from '../../infrastructure/serializers/csv/sup-organization-learner-parser.js';
+import { getDataBuffer } from '../../infrastructure/utils/bufferize/get-data-buffer.js';
 import { AggregateImportError } from '../errors.js';
 import { OrganizationImport } from '../models/OrganizationImport.js';
 
@@ -74,16 +75,4 @@ const replaceSupOrganizationLearners = async function ({
   }
 };
 
-function getDataBuffer(readableStream) {
-  return new Promise((resolve, reject) => {
-    const chunks = [];
-    readableStream.on('data', (data) => {
-      chunks.push(data);
-    });
-    readableStream.on('error', (err) => reject(err));
-    readableStream.once('end', () => {
-      resolve(Buffer.concat(chunks));
-    });
-  });
-}
 export { replaceSupOrganizationLearners };
