+++ api/src/prescription/learner-management/infrastructure/utils/bufferize/get-data-buffer.js
@@ -1,0 +1,14 @@
+function getDataBuffer(readableStream) {
+  return new Promise((resolve, reject) => {
+    const chunks = [];
+    readableStream.on('data', (data) => {
+      chunks.push(data);
+    });
+    readableStream.on('error', (err) => reject(err));
+    readableStream.once('end', () => {
+      resolve(Buffer.concat(chunks));
+    });
+  });
+}
+
+export { getDataBuffer };
